# BE-A2-Database-CRUD/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A small in-memory CRUD API for a to-do list. Built by hand for FlyRank Internship, Backend Track, Week 2, Assignment A1. Data lives in a JavaScript array, so it resets every time the server restarts."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    { "name": "meta", "description": "Information about the API itself" },
    { "name": "tasks", "description": "Create, read, update and delete tasks" }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["meta"],
        "summary": "Describe this API",
        "description": "Returns the API name, version and the list of top-level resources.",
        "responses": {
          "200": {
            "description": "API description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "version": { "type": "string" },
                    "endpoints": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                },
                "example": {
                  "name": "Task API",
                  "version": "1.0",
                  "endpoints": ["/tasks"]
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["meta"],
        "summary": "Liveness probe",
        "description": "Returns 200 as long as the process is running. Monitoring tools poll this.",
        "responses": {
          "200": {
            "description": "The server is alive",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string" } }
                },
                "example": { "status": "ok" }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": ["meta"],
        "summary": "Count tasks by state",
        "description": "The server computes something instead of just storing it.",
        "responses": {
          "200": {
            "description": "Task counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer" },
                    "done": { "type": "integer" },
                    "open": { "type": "integer" }
                  }
                },
                "example": { "total": 3, "done": 1, "open": 2 }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": ["tasks"],
        "summary": "List tasks",
        "description": "Returns every task, optionally filtered, searched and paginated.",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "required": false,
            "description": "Keep only finished (true) or unfinished (false) tasks.",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Keep only tasks whose title contains this text (case-insensitive).",
            "schema": { "type": "string" },
            "example": "milk"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "How many tasks to return at most.",
            "schema": { "type": "integer", "minimum": 1 },
            "example": 2
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "How many tasks to skip before starting.",
            "schema": { "type": "integer", "minimum": 0 },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          },
          "400": {
            "description": "A query parameter had an invalid value",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Query 'limit' must be a positive integer" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["tasks"],
        "summary": "Create a task",
        "description": "The server assigns the id and defaults done to false.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewTask" },
              "example": { "title": "Buy milk" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" },
                "example": { "id": 4, "title": "Buy milk", "done": false }
              }
            }
          },
          "400": {
            "description": "The title was missing, empty or not a string",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Field 'title' is required" }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "The task id.",
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "get": {
        "tags": ["tasks"],
        "summary": "Read one task",
        "responses": {
          "200": {
            "description": "The task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": {
            "description": "No task has that id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["tasks"],
        "summary": "Update a task",
        "description": "Send title, done, or both. Fields you leave out stay unchanged.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskUpdate" },
              "example": { "title": "Buy oat milk", "done": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "The body was empty or a field had the wrong type",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Body must contain 'title' and/or 'done'" }
              }
            }
          },
          "404": {
            "description": "No task has that id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tasks"],
        "summary": "Delete a task",
        "responses": {
          "204": { "description": "Deleted. No body is returned." },
          "404": {
            "description": "No task has that id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "tags": ["tasks"],
        "summary": "Reset to the three seed tasks",
        "description": "Throws away everything in memory and restores the original example tasks. Handy for demos.",
        "responses": {
          "200": {
            "description": "The restored list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "required": ["id", "title", "done"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Assigned by the server.",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "What needs doing.",
            "example": "Buy milk"
          },
          "done": {
            "type": "boolean",
            "description": "Whether the task is finished.",
            "example": false
          }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "description": "Required, and must not be empty.",
            "example": "Buy milk"
          },
          "done": {
            "type": "boolean",
            "description": "Optional. Defaults to false.",
            "example": false
          }
        }
      },
      "TaskUpdate": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "title": { "type": "string", "minLength": 1, "example": "Buy oat milk" },
          "done": { "type": "boolean", "example": true }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "A human-readable explanation of what went wrong.",
            "example": "Task 99 not found"
          }
        }
      }
    }
  }
}